Update gradientdescent.py
Now that the model has been trained, it can accurately predict housing values using the learned parameters. It's no surprise that the predicted values closely match the training values for the same housing. Additionally, the values not included in the prediction align with the anticipated value.n align with the anticipated value.